Replace AuthContext with UserContext
